chore: add AstText component
